+++ pipen/pluginmgr.py
@@ -20,23 +20,23 @@ plugin = Simplug("pipen")
 
 
 @plugin.spec
-def on_setup(plugin_opts: Dict[str, Any]) -> None:
+def on_setup(config: Dict[str, Any]) -> None:
     """Setup for plugins, primarily used for the plugins to
     setup some default configurations.
 
-    This is only called once per python session.
+    This is only called once for all pipelines.
 
     Args:
-        plugin_opts: The plugin configuration dictionary
+        config: The configuration dictionary
+            plugin options should be defined under "plugin_opts"
             One should define a configuration item either with a prefix as
-            the identity for the plugin or a namespace inside the plugin config.
+            the identity for the plugin or a namespace inside the plugin config
     """
 
 
 @plugin.spec
 async def on_init(pipen: "Pipen") -> None:
-    """When the pipeline is initialized, and pipeline-level configuration
-    items are calculated.
+    """When the pipeline is initialized, and default configs are loaded
 
     Args:
         pipen: The Pipen object
@@ -201,6 +201,7 @@ async def on_job_failed(proc: "Proc", job: "Job"):
         job: The job
     """
 
+
 class PipenMainPlugin:
     """The builtin main plugin, used to update the progress bar and
     cache the job"""
@@ -213,7 +214,7 @@ class PipenMainPlugin:
         if sig:  # pragma: no cover
             proc.log(
                 "warning",
-                "Got signal %r, trying a graceful11 shutdown ...",
+                "Got signal %r, trying a graceful shutdown ...",
                 sig.name,
             )